Crate ros_message

Crate for describing ROS messages, and handling MSG and SRV files.

Structs

Duration

ROS representation of duration, with nanosecond precision

FieldInfo

Full description of one field in a msg or srv file.

MessagePath

Path to a ROS message with naming conventions tested.

Msg

A ROS message parsed from a msg file.

Srv

A ROS service parsed from a srv file.

Time

ROS representation of time, with nanosecond precision

Enums

DataType

Enumerates all data types possible in a ROS message.

Error

Enumeration of all errors that can be returned.

FieldCase

Represents all possible variants of a message field

I8Variant

All possible names for a signed 1 byte integer in ROS messages.

U8Variant

All possible names for a signed 1 byte integer in ROS messages.

Value

Represents an arbitrary ROS message or value in it.

Type Aliases

MessageValue

Represents an arbitrary ROS message.

Result

Convenience type for shorter return value syntax of this crate’s errors.