ros_message

Enum Error

pub enum Error {
    InvalidMessagePath {
        name: String,
        reason: String,
    },
    UnsupportedDataType {
        name: String,
        reason: String,
    },
    BadMessageContent(String),
    MessageDependencyMissing {
        package: String,
        name: String,
    },
    BadConstant {
        name: String,
        datatype: String,
        value: String,
    },
}

Enumeration of all errors that can be returned.

Variants

InvalidMessagePath

Message doesnโt have a valid format.

Message names must follow the package_name/MessageName format.

Packages must follow REP 144 rules.

Fields

name: String

Full name of the message we are trying to parse.

reason: String

Reason for the failure.

UnsupportedDataType

Field in the msg or srv file has a name that doesnโt fit any data type category.

Fields

name: String

Full name of the data type we are trying to parse.

reason: String

Reason for the failure.

BadMessageContent(String)

The msg or srv file being parsed has invalid content.

MessageDependencyMissing

Certain operations on a msg or srv file require first handling all messages it depends upon.

For example, to calculate an MD5 sum for a message, you first need to calculate it for all messages it depends upon, and passing them into the calculation call.

Fields

package: String

Package that the message is located in.

name: String

Name of the missing message.

BadConstant

Passed in constant value is not parsable as its data type.

Fields

name: String

Name of the constant.

datatype: String

Type of the invalid value.

value: String

The invalid value provided.