ros_message

Struct Srv

pub struct Srv { /* private fields */ }

A ROS service parsed from a srv file.

Implementations

impl Srv

pub fn new(path: MessagePath, source: impl Into<String>) -> Result<Srv>

Create a service from a passed in path and source.

Errors

Returns an error if there is an error parsing the service source.

Examples

let service = Srv::new(
    "foo/Bar".try_into()?,
    r#"# a comment that is ignored
    Header header
    uint32 a
    byte[16] b
    geometry_msgs/Point[] point
    uint32 FOO=5
    string SOME_TEXT=this is # some text, don't be fooled by the hash
---
    uint32 a
    geometry_msgs/Point[] point
    uint32 FOO=6
    "#,
)?;

assert_eq!(service.path(), &"foo/Bar".try_into()?);
assert_eq!(service.request().fields().len(), 6);
assert_eq!(service.response().fields().len(), 3);

pub fn path(&self) -> &MessagePath

Returns the path of the service.

pub fn source(&self) -> &str

Returns the original source.

pub fn request(&self) -> &Msg

Returns the request message.

pub fn response(&self) -> &Msg

Returns the response message.